+++ src/views/UserSearch/UserSearch.js
@@ -160,6 +160,12 @@ class UserSearch extends React.Component {
   }
 
   componentDidUpdate(prevProps) {
+    const {
+      resources,
+      source,
+      history,
+    } = this.props;
+
     if (
       this.state.searchPending &&
       prevProps.resources.records &&
@@ -168,7 +174,6 @@ class UserSearch extends React.Component {
       this.onSearchComplete(this.props.resources.records);
     }
 
-    const { resources, source, history } = this.props;
     // Open the detail view when there's a single hit
     if (resources.records.records[0] !== prevProps.resources.records.records[0] &&
       source.totalCount() === 1) {
@@ -197,11 +202,14 @@ class UserSearch extends React.Component {
     };
   }
 
-  onSearchComplete = records => {
-    const { intl } = this.props;
+  onSearchComplete = () => {
+    const {
+      intl,
+      actualTotalRecords,
+    } = this.props;
+
     const headerEl = this.resultsPaneTitleRef.current;
-    const resultsCount = get(records, 'other.totalRecords', 0);
-    const hasResults = !!resultsCount;
+    const hasResults = !!actualTotalRecords;
 
     this.setState({ searchPending: false });
 
@@ -209,7 +217,7 @@ class UserSearch extends React.Component {
     this.SRStatusRef.current.sendMessage(intl.formatMessage({
       id: 'ui-users.resultCount',
     }, {
-      count: resultsCount
+      count: actualTotalRecords,
     }));
 
     // Focus the pane header if we have results to minimize tabbing distance
@@ -631,6 +639,13 @@ class UserSearch extends React.Component {
     }
   }
 
+  handleResetAll = (resetAll) => () => {
+    const { onResetActualTotalRecords } = this.props;
+
+    onResetActualTotalRecords();
+    resetAll();
+  }
+
   render() {
     const {
       onComponentWillUnmount,
@@ -646,6 +661,8 @@ class UserSearch extends React.Component {
       stripes: { timezone },
       okapi: { currentUser },
       intl: { formatMessage },
+      actualTotalRecords,
+      hasLoadedActualTotalRecords,
     } = this.props;
     const {
       filterPaneIsVisible,
@@ -670,7 +687,6 @@ class UserSearch extends React.Component {
     const servicePoints = currentUser.servicePoints;
     const patronGroups = (resources.patronGroups || {}).records || [];
     const query = queryGetter ? queryGetter() || {} : {};
-    const count = source ? source.totalCount() : 0;
     const sortOrder = query.sort || '';
     const initialCashDrawerReportValues = {
       format: 'both'
@@ -686,8 +702,9 @@ class UserSearch extends React.Component {
       </div>) : 'no source yet';
 
     let resultPaneSub = <FormattedMessage id="stripes-smart-components.searchCriteria" />;
-    if (source && source.loaded()) {
-      resultPaneSub = <FormattedMessage id="stripes-smart-components.searchResultsCountHeader" values={{ count }} />;
+
+    if (hasLoadedActualTotalRecords) {
+      resultPaneSub = <FormattedMessage id="stripes-smart-components.searchResultsCountHeader" values={{ count: actualTotalRecords }} />;
     }
 
     const resultsFormatter = {
@@ -789,7 +806,7 @@ class UserSearch extends React.Component {
                               id="clickable-reset-all"
                               disabled={!(filterChanged || searchChanged)}
                               fullWidth
-                              onClick={resetAll}
+                              onClick={this.handleResetAll(resetAll)}
                             >
                               <Icon icon="times-circle-solid">
                                 <FormattedMessage id="stripes-smart-components.resetAll" />
@@ -827,7 +844,7 @@ class UserSearch extends React.Component {
                             visibleColumns={visibleColumns}
                             rowUpdater={this.rowUpdater}
                             contentData={users}
-                            totalCount={count}
+                            totalCount={actualTotalRecords}
                             columnMapping={columnMapping}
                             formatter={resultsFormatter}
                             onNeedMoreData={onNeedMoreData}